+++ common/geometry.hpp
@@ -6,12 +6,12 @@
 #include "peseudo_random.hpp"
 
 namespace rt {
-	inline glm::vec3 triNg(const glm::vec3 &v0, const glm::vec3 &v1, const glm::vec3 &v2, bool isback) {
-		glm::vec3 e1 = v1 - v0;
-		glm::vec3 e2 = v2 - v0;
+	inline glm::dvec3 triNg(const glm::dvec3 &v0, const glm::dvec3 &v1, const glm::dvec3 &v2, bool isback) {
+		glm::dvec3 e1 = v1 - v0;
+		glm::dvec3 e2 = v2 - v0;
 		return glm::normalize(isback ? glm::cross(e2, e1) : glm::cross(e1, e2));
 	}
-	inline float triArea(const glm::vec3 &p0, const glm::vec3 &p1, const glm::vec3 &p2) {
+	inline double triArea(const glm::dvec3 &p0, const glm::dvec3 &p1, const glm::dvec3 &p2) {
 		auto va = p0 - p1;
 		auto vb = p2 - p1;
 		return glm::length(glm::cross(va, vb)) * 0.5;
