+++ common/plane.hpp
@@ -7,18 +7,18 @@ namespace rt {
 	// Plane equation
 	// d = dot(n, p) for a given point p on the plane
 	struct Plane {
-		glm::vec3 n;
-		float d = 0.0f;
+		glm::dvec3 n;
+		double d = 0.0;
 	};
-	inline Plane plane_from(const glm::vec3 &n, const glm::vec3 &p) {
+	inline Plane plane_from(const glm::dvec3 &n, const glm::dvec3 &p) {
 		Plane plane;
 		plane.n = n;
 		plane.d = glm::dot(plane.n, p);
 		return plane;
 	}
 
-	inline bool intersect_ray_plane(const glm::vec3 &o, const glm::vec3 &d, const Plane &plane, float *tmin) {
-		float eps = 1.0e-5f;
+	inline bool intersect_ray_plane(const glm::dvec3 &o, const glm::dvec3 &d, const Plane &plane, double *tmin) {
+		double eps = 1.0e-5f;
 		auto denom = glm::dot(plane.n, d);
 		if (std::fabs(denom) < eps) {
 			return false;
